Move test helpers to Feature namespace
They are used only there.
