+++ tests/Feature/ConfigTest.php
@@ -3,7 +3,7 @@
 namespace tests\Feature;
 
 use tests\Feature\Helper\WatcherTestCase;
-use tests\Helper\Filesystem;
+use tests\Feature\Helper\Filesystem;
 use tests\Helper\WatcherRunner;
 
 final class ConfigTest extends WatcherTestCase
