+++ tests/Feature/Helper/WithFilesystem.php
@@ -1,6 +1,6 @@
 <?php declare(strict_types=1);
 
-namespace tests\Helper;
+namespace tests\Feature\Helper;
 
 trait WithFilesystem
 {
